addPermission

+++ src/app/ABMDiscounts/page.tsx
@@ -8,9 +8,10 @@ import Table from "@/Components/Atoms/Table/Table";
 import style from './ABMDiscounts.module.scss'
 import { useEffect, useState } from "react";
 import discountServices from '../../Services/discountServices'
-import Swal from 'sweetalert2'
 import GuardLogin from "@/utils/guardLogin";
 import Loader from "@/Components/Organism/loaderScreen/loader";
+import AlertServices from "@/utils/AlertServices";
+import sessionServices from "@/Services/sessionServices";
 
 type IAllDiscounts = {
     id: number,
@@ -46,27 +47,39 @@ const ABMDiscount = () => {
     }
 
     async function deleteElementFunction(index:number) {
-        const elementToDelete = discountsAllData[index]
-        const newDiscountsAllData = discountsAllData.filter((obj) => obj.id !== elementToDelete.id);
-        const dataPricesInTable = formatDiscountToTable(newDiscountsAllData)
-        setDiscountsAllData(newDiscountsAllData)
-        setDiscountsData(dataPricesInTable)
-
-        await discountServices.deleteDiscount(elementToDelete.id)
-        Swal.fire({
-            title: 'El Descuento ha sido eliminado',
-            icon: 'success',
-            confirmButtonText: 'Cerrar',
-          })
+        if(sessionServices.isAdmin()) {
+            const elementToDelete = discountsAllData[index]
+            const newDiscountsAllData = discountsAllData.filter((obj) => obj.id !== elementToDelete.id);
+            const dataPricesInTable = formatDiscountToTable(newDiscountsAllData)
+            setDiscountsAllData(newDiscountsAllData)
+            setDiscountsData(dataPricesInTable)
+    
+            await discountServices.deleteDiscount(elementToDelete.id)
+            AlertServices.renderAlert(
+                'El Descuento ha sido eliminado',
+                '',
+                'success'
+            )
+            return
+        }
+        AlertServices.renderAlertPermission()
     } 
 
     const openModalEditFunction = () => {
-        setOpenModalEdit(true)
+        if(sessionServices.isAdmin()) {
+            setOpenModalEdit(true)
+            return
+        }
+        AlertServices.renderAlertPermission();
     }
 
     const openModalCreateFunction = () => {
-        setFullDiscountToEdit({ id: '', name: '', percentage: ''})
-        setOpenModalCreate(true)
+        if(sessionServices.isAdmin()) {
+            setFullDiscountToEdit({ id: '', name: '', percentage: ''})
+            setOpenModalCreate(true)
+            return
+        }
+        AlertServices.renderAlertPermission();
     }
